refactor: extract ComputeDownloadManager to separate file, gate behind GOW3 serials
Move all compute→texture download logic into its own compilation unit
(vk_compute_download.h/.cpp) using PIMPL to minimize header intrusion.
The only additions to vk_rasterizer.h are a forward-declared unique_ptr
and one wrapper method declaration.

Feature activation is now gated by game serial — only enabled for the
three God of War III Remastered language versions (CUSA01623/CUSA01715/
CUSA01740). Activation is logged at INFO level.

Changes:
- New: vk_compute_download.h (PIMPL header, 40 lines)
- New: vk_compute_download.cpp (full implementation, ~230 lines)
- vk_rasterizer.h: -PendingComputeDownload struct, +unique_ptr to manager
- vk_rasterizer.cpp: SyncComputeStorageImages and ResolvePending now
  thin wrappers delegating to ComputeDownloadManager

[Claude Code, deepseek-v4-pro[1m]]
